+++ Common/src/graphics/svg/element/shape/path/CubicTo.java
@@ -233,12 +233,12 @@ public class CubicTo extends PathOp
 	{
 		if (absolute())
 		{
-			path.curveTo(x0+x1, y0+y1, x0+x2, y0+y2, x0+x, y0+y);
+			path.curveTo((float) (x0+x1), (float) (y0+y1), (float) (x0+x2), (float) (y0+y2), (float) (x0+x), (float) (y0+y));
 		}
 		else
 		{
 			final Point2D pt = path.getCurrentPoint();
-			path.curveTo(pt.getX()+x1, pt.getY()+y1, pt.getX()+x2, pt.getY()+y2, pt.getX()+x, pt.getY()+y);
+			path.curveTo((float) (pt.getX()+x1), (float) (pt.getY()+y1), (float) (pt.getX()+x2), (float) (pt.getY()+y2), (float) (pt.getX()+x), (float) (pt.getY()+y));
 		}
 //		return new Point2D.Double(x2, y2);
 	}
